refactor: removal of cyclic dependencies between index/txindex, validation and index/base translation units
