+++ src/Makefile.am
@@ -185,6 +185,7 @@ BITCOIN_CORE_H = \
   utilmoneystr.h \
   utiltime.h \
   validation.h \
+  validation_globals.h \
   validationinterface.h \
   versionbits.h \
   walletinitinterface.h \
@@ -256,6 +257,7 @@ libbitcoin_server_a_SOURCES = \
   txmempool.cpp \
   ui_interface.cpp \
   validation.cpp \
+  validation_globals.cpp \
   validationinterface.cpp \
   versionbits.cpp \
   $(BITCOIN_CORE_H)
