+++ src/validation_globals.h
@@ -1,0 +1,114 @@
+#ifndef BITCOIN_VALIDATION_GLOBALS_H
+#define BITCOIN_VALIDATION_GLOBALS_H
+
+#include <chain.h>
+#include <util.h>
+#include <protocol.h>
+
+#include <unordered_map>
+
+class CCoinsViewDB;
+class CBlockTreeDB;
+class CCoinsViewCache;
+class CBlockPolicyEstimator; 
+class CTxMemPool;
+class CFeeRate;
+
+struct BlockHasher
+{
+    size_t operator()(const uint256& hash) const { return hash.GetCheapHash(); }
+};
+
+extern CScript COINBASE_FLAGS;
+extern CCriticalSection cs_main;
+extern CBlockPolicyEstimator feeEstimator;
+extern CTxMemPool mempool;
+extern std::atomic_bool g_is_mempool_loaded;
+typedef std::unordered_map<uint256, CBlockIndex*, BlockHasher> BlockMap;
+extern BlockMap& mapBlockIndex;
+extern uint64_t nLastBlockTx;
+extern uint64_t nLastBlockWeight;
+extern const std::string strMessageMagic;
+extern CWaitableCriticalSection g_best_block_mutex;
+extern CConditionVariable g_best_block_cv;
+extern uint256 g_best_block;
+extern std::atomic_bool fImporting;
+extern std::atomic_bool fReindex;
+extern int nScriptCheckThreads;
+extern bool fIsBareMultisigStd;
+extern bool fRequireStandard;
+extern bool fCheckBlockIndex;
+extern bool fCheckpointsEnabled;
+extern size_t nCoinCacheUsage;
+
+/** Block hash whose ancestors we will assume to have valid scripts without checking them. */
+extern uint256 hashAssumeValid;
+
+/** Minimum work we will assume exists on some valid chain. */
+extern arith_uint256 nMinimumChainWork;
+
+/** Best header we've seen so far (used for getheaders queries' starting points). */
+extern CBlockIndex *pindexBestHeader;
+
+/** Minimum disk space required - used in CheckDiskSpace() */
+static const uint64_t nMinDiskSpace = 52428800;
+
+/** Pruning-related variables and constants */
+/** True if any block files have ever been pruned. */
+extern bool fHavePruned;
+/** True if we're running in -prune mode. */
+extern bool fPruneMode;
+/** Number of MiB of block files that we're trying to stay below. */
+extern uint64_t nPruneTarget;
+/** Block files containing a block-height within MIN_BLOCKS_TO_KEEP of chainActive.Tip() will not be pruned. */
+static const unsigned int MIN_BLOCKS_TO_KEEP = 288;
+/** Minimum blocks required to signal NODE_NETWORK_LIMITED */
+static const unsigned int NODE_NETWORK_LIMITED_MIN_BLOCKS = 288;
+
+static const signed int DEFAULT_CHECKBLOCKS = 6;
+static const unsigned int DEFAULT_CHECKLEVEL = 3;
+
+
+
+/** A fee rate smaller than this is considered zero fee (for relaying, mining and transaction creation) */
+extern CFeeRate minRelayTxFee;
+/** Absolute maximum transaction fee (in satoshis) used by wallet and mempool (rejects high fee in sendrawtransaction) */
+extern CAmount maxTxFee;
+/** If the tip is older than this (in seconds), the node is considered to be in initial block download. */
+extern int64_t nMaxTipAge;
+extern bool fEnableReplacement;
+/** Functions for disk access for blocks */
+bool ReadBlockFromDisk(CBlock& block, const CDiskBlockPos& pos, const Consensus::Params& consensusParams);
+bool ReadBlockFromDisk(CBlock& block, const CBlockIndex* pindex, const Consensus::Params& consensusParams);
+bool ReadRawBlockFromDisk(std::vector<uint8_t>& block, const CDiskBlockPos& pos, const CMessageHeader::MessageStartChars& message_start);
+bool ReadRawBlockFromDisk(std::vector<uint8_t>& block, const CBlockIndex* pindex, const CMessageHeader::MessageStartChars& message_start);
+
+/** Find the last common block between the parameter chain and a locator. */
+CBlockIndex* FindForkInGlobalIndex(const CChain& chain, const CBlockLocator& locator);
+
+fs::path GetBlockPosFilename(const CDiskBlockPos &pos, const char *prefix);
+
+FILE* OpenDiskFile(const CDiskBlockPos &pos, const char *prefix, bool fReadOnly);
+
+/** Open a block file (blk?????.dat) */
+FILE* OpenBlockFile(const CDiskBlockPos &pos, bool fReadOnly = false);
+
+/** The currently-connected chain of blocks (protected by cs_main). */
+extern CChain& chainActive;
+
+/** Global variable that points to the coins database (protected by cs_main) */
+extern std::unique_ptr<CCoinsViewDB> pcoinsdbview;
+
+/** Global variable that points to the active CCoinsView (protected by cs_main) */
+extern std::unique_ptr<CCoinsViewCache> pcoinsTip;
+
+/** Global variable that points to the active block tree (protected by cs_main) */
+extern std::unique_ptr<CBlockTreeDB> pblocktree;
+inline CBlockIndex* LookupBlockIndex(const uint256& hash)
+{
+    AssertLockHeld(cs_main);
+    BlockMap::const_iterator it = mapBlockIndex.find(hash);
+    return it == mapBlockIndex.end() ? nullptr : it->second;
+}
+
+#endif // BITCOIN_VALIDATION_GLOBALS_H
